fix(terraform_module): do not reuse terraform_module_workdir
An error would occur when the role was used twice in the same playbook.
The variable terraform_module_workdir would be set by the first run, and
the subsequent run would not create a new temporary directory and attempt
to reuse the previous (deleted) directory. As that directory did not
exist, this would fail.

This commit adds a new variable _terraform_module_workdir that allows for
the terraform_module_workdir variable to remain unset if originally unset.

Fixes: #1