Add override for `terraform_state_aws_region`
The current method for getting the aws region using the aws cli does
not work when you use granted [1] for assuming roles.

This patch allows the caller to just define and pass in the region for
the state bucket/table. It maintains backwards compatibility by
attempting to figure out the region if none is passed.

[1]: https://www.granted.dev/